requirements.txt

python-telegram-bot==20.6
flask
gspread
oauth2client
